third_party/WebKit/LayoutTests/http/tests/security/isolatedWorld/cross-origin-xhr-expected.txt

CONSOLE ERROR: Failed to load http://localhost:8000/security/isolatedWorld/resources/cross-origin-xhr.txt: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Failed to load http://localhost:8000/security/isolatedWorld/resources/cross-origin-xhr.txt: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Failed to load http://localhost:8000/security/isolatedWorld/resources/cross-origin-xhr.txt: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Tests that isolated worlds can have XHRs run in a different origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


XHR from main world
XHR completed with status 0
XHR from isolated world with unchanged security origin
XHR completed with status 0
XHR from isolated world with same security origin as XHR target
XHR completed with status 200
XHR from isolated world with security origin that is whitelisted for XHR target
XHR completed with status 200
XHR from main world is not affected by the isolated world origin or whitelist
XHR completed with status 0
PASS successfullyParsed is true

TEST COMPLETE
